hermit::arch
Trait PageSize

pub trait PageSize:
    Copy
    + Eq
    + PartialOrd
    + Ord
    + Sealed {
    const SIZE: u64;
    const DEBUG_STR: &'static str;
}

Trait for abstracting over the three possible page sizes on x86_64, 4KiB, 2MiB, 1GiB.

Required Associated Constants

const SIZE: u64

The page size in bytes.

const DEBUG_STR: &'static str

A string representation of the page size for debug output.

Object Safety

This trait is not object safe.