hermit
Module errno

Constants

E2BIG

Argument list too long

EACCES

Permission denied

EADDRINUSE

Address already in use

EADDRNOTAVAIL

Cannot assign requested address

EADV

Advertise error

EAFNOSUPPORT

Address family not supported by protocol

EAGAIN

Try again

EALREADY

Operation already in progress

EBADE

Invalid exchange

EBADF

Bad file number

EBADFD

File descriptor in bad state

EBADMSG

Not a data message

EBADR

Invalid request descriptor

EBADRQC

Invalid request code

EBADSLT

Invalid slot

EBFONT

Bad font file format

EBUSY

Device or resource busy

ECANCELED

Operation Canceled

ECHILD

No child processes

ECHRNG

Channel number out of range

ECOMM

Communication error on send

ECONNABORTED

Software caused connection abort

ECONNREFUSED

Connection refused

ECONNRESET

Connection reset by peer

EDEADLK

Resource deadlock would occur

EDEADLOCK

EDESTADDRREQ

Destination address required

EDOM

Math argument out of domain of func

EDOTDOT

RFS specific error

EDQUOT

Quota exceeded

EEXIST

File exists

EFAULT

Bad address

EFBIG

File too large

EHOSTDOWN

Host is down

EHOSTUNREACH

No route to host

EHWPOISON

Robust mutexes: Memory page has hardware error

EIDRM

Identifier removed

EILSEQ

Illegal byte sequence

EINPROGRESS

Operation now in progress

EINTR

Interrupted system call

EINVAL

Invalid argument

EIO

I/O error

EISCONN

Transport endpoint is already connected

EISDIR

Is a directory

EISNAM

Is a named type file

EKEYEXPIRED

Key has expired

EKEYREJECTED

Key was rejected by service

EKEYREVOKED

Key has been revoked

EL2HLT

Level 2 halted

EL2NSYNC

Level 2 not synchronized

EL3HLT

Level 3 halted

EL3RST

Level 3 reset

ELIBACC

Can not access a needed shared library

ELIBBAD

Accessing a corrupted shared library

ELIBEXEC

Cannot exec a shared library directly

ELIBMAX

Attempting to link in too many shared libraries

ELIBSCN

.lib section in a.out corrupted

ELNRNG

Link number out of range

ELOOP

Too many symbolic links encountered

EMEDIUMTYPE

Wrong medium type

EMFILE

Too many open files

EMLINK

Too many links

EMSGSIZE

Message too long

EMULTIHOP

Multihop attempted

ENAMETOOLONG

File name too long

ENAVAIL

No XENIX semaphores available

ENETDOWN

Network is down

ENETRESET

Network dropped connection because of reset

ENETUNREACH

Network is unreachable

ENFILE

File table overflow

ENOANO

No anode

ENOBUFS

No buffer space available

ENOCSI

No CSI structure available

ENODATA

No data available

ENODEV

No such device

ENOENT

No such file or directory

ENOEXEC

Exec format error

ENOKEY

Required key not available

ENOLCK

No record locks available

ENOLINK

Link has been severed

ENOMEDIUM

No medium found

ENOMEM

Out of memory

ENOMSG

No message of desired type

ENONET

Machine is not on the network

ENOPKG

Package not installed

ENOPROTOOPT

Protocol not available

ENOSPC

No space left on device

ENOSR

Out of streams resources

ENOSTR

Device not a stream

ENOSYS

Function not implemented

ENOTBLK

Block device required

ENOTCONN

Transport endpoint is not connected

ENOTDIR

Not a directory

ENOTEMPTY

Directory not empty

ENOTNAM

Not a XENIX named type file

ENOTRECOVERABLE

Robust mutexes: State not recoverable

ENOTSOCK

Socket operation on non-socket

ENOTTY

Not a typewriter

ENOTUNIQ

Name not unique on network

ENXIO

No such device or address

EOPNOTSUPP

Operation not supported on transport endpoint

EOVERFLOW

Value too large for defined data type

EOWNERDEAD

Robust mutexes: Owner died

EPERM

Operation not permitted

EPFNOSUPPORT

Protocol family not supported

EPIPE

Broken pipe

EPROTO

Protocol error

EPROTONOSUPPORT

Protocol not supported

EPROTOTYPE

Protocol wrong type for socket

ERANGE

Math result not representable

EREMCHG

Remote address changed

EREMOTE

Object is remote

EREMOTEIO

Remote I/O error

ERESTART

Interrupted system call should be restarted

ERFKILL

Robust mutexes: Operation not possible due to RF-kill

EROFS

Read-only file system

ESHUTDOWN

Cannot send after transport endpoint shutdown

ESOCKTNOSUPPORT

Socket type not supported

ESPIPE

Illegal seek

ESRCH

No such process

ESRMNT

Srmount error

ESTALE

Stale file handle

ESTRPIPE

Streams pipe error

ETIME

Timer expired

ETIMEDOUT

Connection timed out

ETOOMANYREFS

Too many references: cannot splice

ETXTBSY

Text file busy

EUCLEAN

Structure needs cleaning

EUNATCH

Protocol driver not attached

EUSERS

Too many users

EWOULDBLOCK

Operation would block

EXDEV

Cross-device link

EXFULL

Exchange full

Functions

sys_errno

Get the error number from the thread local storage

sys_get_errno

Get the error number from the thread local storage