Module hermit::fd

Structs

AccessPermission

EventFlags

OpenOption

Options for opening files

PollEvent

PollFd

Enums

IoError

Functions

eventfd

eventfd creates an linux-like “eventfd object” that can be used as an event wait/notify mechanism by user-space applications, and by the kernel to notify user-space applications of events. The object contains an unsigned 64-bit integer counter that is maintained by the kernel. This counter is initialized with the value specified in the argument initval.

poll

The unix-like poll waits for one of a set of file descriptors to become ready to perform I/O. The set of file descriptors to be monitored is specified in the fds argument, which is an array of structs of PollFd.