hermit::fd

Function poll

pub fn poll(fds: &mut [PollFd], timeout: Option<Duration>) -> Result<u64>

Wait for some event on a file descriptor.

The unix-like poll waits for one of a set of file descriptors to become ready to perform I/O. The set of file descriptors to be monitored is specified in the fds argument, which is an array of structs of PollFd.