Struct hermit::fs::Filesystem

pub struct Filesystem { /* fields omitted */ }

Implementations

impl Filesystem

pub const fn new() -> Self

pub fn open(&mut self, path: &str, perms: FilePerms) -> Result<u64, FileError>

Tries to open file at given path (/MOUNTPOINT/internal-path). Looks up MOUNTPOINT in mounted dirs, passes internal-path to filesystem backend Returns the file descriptor of the newly opened file, or an error on failure

pub fn close(&mut self, fd: u64)

pub fn unlink(&mut self, path: &str) -> Result<(), FileError>

Unlinks a file given by path

pub fn mount(
&mut self,
mntpath: &str,
mntobj: Box<dyn PosixFileSystem + Send>
) -> Result<(), ()>

Create new backing-fs at mountpoint mntpath

pub fn fd_op(&mut self, fd: u64, f: impl FnOnce(&mut Box<dyn PosixFile + Send>))

Run closure on file referenced by file descriptor.