Module hermit::irq

Structs

ExceptionStackFrame
Represents the exception stack frame pushed by the CPU on exception entry.

IrqStatistics

Functions

add_irq_name

disable
Disable Interrupts

enable
Enable Interrupts

enable_and_wait
Enable Interrupts and wait for the next interrupt (HLT instruction) According to https://lists.freebsd.org/pipermail/freebsd-current/2004-June/029369.html, this exact sequence of assembly instructions is guaranteed to be atomic. This is important, because another CPU could call wakeup_core right when we decide to wait for the next interrupt.

get_irq_name

install

irq_install_handler

nested_disable
Disable IRQs (nested)

nested_enable
Enable IRQs (nested)