Module hermit::kernel::processor

Structs

FPUState

XSaveAVXState

XSaveBndcsr

XSaveBndregs

XSaveHeader

XSaveLWPState

XSave Area for AMD Lightweight Profiling. Refer to AMD Lightweight Profiling Specification (Publication No. 43724), Figure 7-1.

XSaveLegacyRegion

Functions

configure

detect_features

detect_frequency

generate_random_number32

generate_random_number64

get_frequency

get_linear_address_bits

get_physical_address_bits

get_timer_ticks

get_timestamp

halt

The halt function stops the processor until the next interrupt arrives

print_information

readfs

readgs

run_on_hypervisor

shutdown

Shutdown the system

supports_1gib_pages

supports_avx

supports_fsgs

supports_tsc_deadline

supports_x2apic

supports_xsave

udelay

Delay execution by the given number of microseconds using busy-waiting.

writefs

writegs