Module hermit::switch

Functions

kernel_function0
Executes f on the kernel stack.

kernel_function1
Executes f on the kernel stack.

kernel_function2
Executes f on the kernel stack.

kernel_function3
Executes f on the kernel stack.

kernel_function4
Executes f on the kernel stack.

kernel_function5
Executes f on the kernel stack.

kernel_function6
Executes f on the kernel stack.

switch_to_fpu_owner⚠
Performa a context switch to an idle task or a task, which alread is owner of the FPU.

switch_to_task⚠