Module hermit::syscalls

Modules

socket

Structs

Dirent64

SpinlockContainer

SpinlockIrqSaveContainer

Functions

sys_abort

sys_add_queue⚠

sys_block_current_task

Set the current task state to blocked

sys_block_current_task_with_timeout

Set the current task state to blocked

sys_clock_getres⚠

Finds the resolution (or precision) of a clock.

sys_clock_gettime⚠

Get the current time of a clock.

sys_clock_nanosleep⚠

Sleep a clock for a specified number of nanoseconds.

sys_clock_settime⚠

sys_close

sys_destroy_queue⚠

sys_dup

sys_eventfd

sys_exit

sys_fcntl

manipulate file descriptor

sys_free⚠

Interface to deallocate a memory region from the system heap

sys_fstat⚠

sys_futex_wait⚠

Like synch::futex_wait, but does extra sanity checks and takes a timespec.

sys_futex_wake⚠

Like synch::futex_wake, but does extra sanity checks.

sys_get_priority

Determine the priority of the current thread

sys_get_processor_count

Returns the number of processors currently online.

sys_get_processor_frequency

Returns the processor frequency in MHz.

sys_getdents64⚠

sys_getpagesize

sys_getpid

sys_gettimeofday⚠

Get the system’s clock time.

sys_image_start_addr

sys_init_queue⚠

sys_ioctl⚠

sys_join

sys_lseek

sys_lstat⚠

sys_malloc

Interface to allocate memory from system heap

sys_mkdir⚠

sys_msleep

sys_nanosleep⚠

sys_notify⚠

sys_open⚠

sys_opendir⚠

sys_poll⚠

sys_rand

The function computes a sequence of pseudo-random integers in the range of 0 to RAND_MAX

sys_read⚠

sys_read_entropy⚠

Fill len bytes in buf with cryptographically secure random data.

sys_realloc⚠

Shrink or grow a block of memory to the given new_size. The block is described by the given ptr pointer and layout. If this returns a non-null pointer, then ownership of the memory block referenced by ptr has been transferred to this allocator. The memory may or may not have been deallocated, and should be considered unusable (unless of course it was transferred back to the caller again via the return value of this method). The new memory block is allocated with layout, but with the size updated to new_size. If this method returns null, then ownership of the memory block has not been transferred to this allocator, and the contents of the memory block are unaltered.

sys_rmdir⚠

sys_secure_rand32⚠

Create a cryptographicly secure 32bit random number with the support of the underlying hardware. If the required hardware isn’t available, the function returns -1.

sys_secure_rand64⚠

Create a cryptographicly secure 64bit random number with the support of the underlying hardware. If the required hardware isn’t available, the function returns -1.

sys_sem_cancelablewait⚠

sys_sem_destroy⚠

Destroy and deallocate a semaphore.

sys_sem_init⚠

Create a new, unnamed semaphore.

sys_sem_post⚠

Release a semaphore.

sys_sem_timedwait⚠

Try to acquire a lock on a semaphore, blocking for a given amount of milliseconds.

sys_sem_trywait⚠

Try to acquire a lock on a semaphore.

sys_set_current_task_priority

Set priority of the current thread

sys_set_priority

Set priority of the thread with the identifier id

sys_setitimer⚠

sys_shutdown

sys_spawn⚠

sys_spawn2⚠

sys_spinlock_destroy⚠

sys_spinlock_init⚠

sys_spinlock_irqsave_destroy⚠

sys_spinlock_irqsave_init⚠

sys_spinlock_irqsave_lock⚠

sys_spinlock_irqsave_unlock⚠

sys_spinlock_lock⚠

sys_spinlock_unlock⚠

sys_srand

The function sets its argument as the seed for a new sequence of pseudo-random numbers to be returned by rand()

sys_stat⚠

sys_thread_exit

sys_unlink⚠

sys_usleep

sys_wait⚠

sys_wakeup_task

Wake up the task with the identifier id

sys_write⚠

sys_yield

Type Aliases

Tid