Module hermit::x86_64::kernel

Modules

acpi

apic

fuse

gdt

idt

irq

pci

percore

pic

pit

processor

scheduler

Architecture dependent interface to initialize a task

serial

switch

systemtime

Structs

BootInfo

Functions

application_processor_init

Application Processor initialization

boot_application_processors

Boots all available Application Processors on bare-metal or QEMU. Called after the Boot Processor has been fully initialized along with its scheduler.

boot_processor_init

Real Boot Processor initialization as soon as we have put the first Welcome message on the screen.

get_base_address

get_cmdline

get_cmdsize

get_image_size

get_limit

get_mbinfo

get_processor_count

get_tls_align

get_tls_filesz

get_tls_memsz

get_tls_start

has_ipdevice

is_single_kernel

Whether HermitCore is running alone (true) or side-by-side to Linux in Multi-Kernel mode (false).

is_uhyve

Whether HermitCore is running under the “uhyve” hypervisor.

is_uhyve_with_pci

message_output_init

Earliest initialization function called by the Boot Processor.

output_message_buf

output_message_byte

print_statistics

sys_uhyve_get_gateway

sys_uhyve_get_ip

sys_uhyve_get_mask