Module hermit::x86_64::kernel::virtio

Modules

consts

Structs

VirtioNotification

Virtq

virtio_pci_common_cfg

virtio_pci_notify_cap
4.1.4.4 Notification structure layout The notification location is found using the VIRTIO_PCI_CAP_NOTIFY_CFG capability. This capability is immediately followed by an additional field, notify_off_multiplier

virtq_desc_raw

Functions

init_virtio_device

map_virtiocap
Scans pci-capabilities for a virtio-capability of type virtiocaptype. When found, maps it into memory and returns virtual address, else None