Crate hermit

First version is derived and adapted for Hermit from Philipp Oppermann’s excellent series of blog posts (http://blog.phil-opp.com/) and Eric Kidd’s toy OS (https://github.com/emk/toyos-rs).

Re-exports

pub use crate::fs::create_file;

Modules

arch

Architecture-specific architecture abstraction.

console

errno

fd

fs

io

scheduler

syscalls

time

Macros

dbg

Prints and returns the value of a given expression for quick and dirty debugging.

panic_println

Emergency output.

print

Prints to the standard output.

println

Prints to the standard output, with a newline.

Constants

DEFAULT_STACK_SIZE