Crate hermit

First version is derived and adapted for HermitCore from Philipp Oppermann’s excellent series of blog posts (http://blog.phil-opp.com/) and Eric Kidd’s toy OS (https://github.com/emk/toyos-rs).

Modules

errno

Macros

dbg

Prints and returns the value of a given expression for quick and dirty debugging.

print

Print formatted text to our console.

println

Print formatted text to our console, followed by a newline.

Structs

SpinlockContainer

SpinlockIrqSaveContainer

itimerval

timespec

timeval

Functions

sys_abort

sys_add_queue⚠

sys_block_current_task

Set the current task state to blocked

sys_block_current_task_with_timeout

Set the current task state to blocked

sys_clock_getres

sys_clock_gettime

sys_clock_nanosleep

sys_clock_settime

sys_close

sys_destroy_queue⚠

sys_exit

sys_free

sys_free_tx_buffer

sys_get_mac_address

sys_get_mtu

sys_get_priority

Determine the priority of the current thread

sys_get_processor_count

Returns the number of processors currently online.

sys_get_processor_frequency

Returns the processor frequency in MHz.

sys_get_tx_buffer

sys_getpagesize

sys_getpid

sys_getprio

sys_gettimeofday

sys_init_queue⚠

sys_join

sys_lseek

sys_malloc

sys_msleep

sys_nanosleep

sys_netwait

sys_notify⚠

sys_open

sys_rand

The function computes a sequence of pseudo-random integers in the range of 0 to RAND_MAX

sys_read

sys_realloc

sys_receive_rx_buffer

sys_recmutex_destroy

sys_recmutex_init

sys_recmutex_lock

sys_recmutex_unlock

sys_rx_buffer_consumed

sys_secure_rand32

Create a cryptographicly secure 32bit random number with the support of the underlying hardware. If the required hardware isn’t available, the function returns None.

sys_secure_rand64

Create a cryptographicly secure 64bit random number with the support of the underlying hardware. If the required hardware isn’t available, the function returns None.

sys_sem_cancelablewait

sys_sem_destroy

sys_sem_init

sys_sem_post

sys_sem_timedwait

sys_sem_trywait

sys_send_tx_buffer

sys_set_network_polling_mode

sys_setitimer

sys_setprio

sys_shutdown

sys_spawn

sys_spawn2

sys_spinlock_destroy

sys_spinlock_init

sys_spinlock_irqsave_destroy

sys_spinlock_irqsave_init

sys_spinlock_irqsave_lock

sys_spinlock_irqsave_unlock

sys_spinlock_lock

sys_spinlock_unlock

sys_srand

The function sets its argument as the seed for a new sequence of pseudo-random numbers to be returned by rand()

sys_stat

sys_thread_exit

sys_unlink

sys_usleep

sys_wait⚠

sys_wakeup_task

Wake up the task with the identifier id

sys_write

sys_yield

Type Definitions

Tid