hermit

Module fs

Structs

DirectoryEntry

File

FileAttr

Metadata

Enums

FileType

SeekWhence

Functions

create_dir

Creates a new, empty directory at the provided path

create_file

file_attributes

metadata

Given a path, query the file system to get information about a file, directory, etc.

open

read_lstat

read_stat

readdir

Returns an vector with all the entries within a directory.

remove_dir

Removes an empty directory.

unlink