Module hermit::kernel::fuse

Structs

Cmd

Fuse

Rsp

fuse_attr

fuse_create_in

fuse_create_out

fuse_entry_out

fuse_in_header

fuse_init_in

fuse_init_out

fuse_lookup_in

fuse_open_in

fuse_open_out

fuse_out_header

fuse_read_in

fuse_read_out

fuse_release_in

fuse_release_out

fuse_unlink_in

fuse_unlink_out

fuse_write_in

fuse_write_out

Enums

Opcode

Traits

FuseIn
Marker trait, which signals that a struct is a valid Fuse command. Struct has to be repr(C)!

FuseInterface

FuseOut
Marker trait, which signals that a struct is a valid Fuse response. Struct has to be repr(C)!

Functions

create_create

create_in_header

create_init

create_lookup

create_open

create_read

create_release

create_unlink

create_write