Module hermit::x86_64::mm::paging

Structs

PageTableEntry

An entry in either table (PML4, PDPT, PD, PT)

PageTableEntryFlags

Possible flags for an entry in either table (PML4, PDPT, PD, PT)

Enums

BasePageSize

A 4 KiB page mapped in the PT.

HugePageSize

A 1 GiB page mapped in the PDPT.

LargePageSize

A 2 MiB page mapped in the PD.

Traits

PageSize

A generic interface to support all possible page sizes.

Functions

get_application_page_size

get_page_table_entry

get_physical_address

identity_map

init

init_page_tables

map

page_fault_handler

unmap

virt_to_phys

virtual_to_physical

Translate a virtual memory address to a physical one.