Trait hermit::x86_64::mm::paging::PageSize

pub trait PageSize: Copy {
    const SIZE: usize;
    const MAP_LEVEL: usize;
    const MAP_EXTRA_FLAG: PageTableEntryFlags;
}

A generic interface to support all possible page sizes.

This is defined as a subtrait of Copy to enable #[derive(Clone, Copy)] for Page. Currently, deriving implementations for these traits only works if all dependent types implement it as well.

Associated Constants

const SIZE: usize

The page size in bytes.

const MAP_LEVEL: usize

The page table level at which a page of this size is mapped (from 0 for PT through 3 for PML4). Implemented as a numeric value to enable numeric comparisons.

const MAP_EXTRA_FLAG: PageTableEntryFlags

Any extra flag that needs to be set to map a page of this size. For example: PageTableEntryFlags::HUGE_PAGE